Fix restrict pushes breaking change in v6.0.0 Terraform provider for Github
Looks like this PR: https://github.com/integrations/terraform-provider-github/pull/2045 introduced a breaking change in the v6.0.0 release of the Terraform provider for GitHub. No note of breaking changes on the release page but it was a major release so breaking changes are likely. https://github.com/integrations/terraform-provider-github/releases/tag/v6.0.0